+++ Modules/Microsoft365DSC/Modules/M365DSCDRGUtil.psm1
@@ -292,8 +292,9 @@ function Get-M365DSCDRGComplexTypeToString
     {
         $currentProperty = @()
         $IndentLevel++
-        foreach ($item in $ComplexObject)
+        for($i = 0; $i -lt $ComplexObject.Count; $i++)
         {
+            $item = $ComplexObject[$i]
             $splat = @{
                 'ComplexObject'   = $item
                 'CIMInstanceName' = $CIMInstanceName
@@ -370,7 +371,6 @@ function Get-M365DSCDRGComplexTypeToString
                 {
                     $hashPropertyType = ([Array]($ComplexTypeMapping | Where-Object -FilterScript { $_.Name -eq $key }).CimInstanceName)[0]
                     $hashProperty = $itemValue
-                    #$currentProperty += "`r`n"
                 }
                 else
                 {
@@ -387,7 +387,7 @@ function Get-M365DSCDRGComplexTypeToString
                     if ($ComplexObject.$key.Count -gt 0)
                     {
                         $currentProperty += $indent + $key + ' = '
-                        $currentProperty += '@('
+                        $currentProperty += "@("
                     }
                 }
 
@@ -413,11 +413,13 @@ function Get-M365DSCDRGComplexTypeToString
                         }
                         if ($i -ne 0)
                         {
-                            # Remove the line break at the start because every item contains a trailing line break
-                            # which would lead to two line breaks between each item
                             $nestedPropertyString = $nestedPropertyString.Substring(2)
                         }
                         $currentProperty += $nestedPropertyString
+                        if (-not $currentProperty.EndsWith("`r`n"))
+                        {
+                            $currentProperty += "`r`n"
+                        }
                     }
                     $IndentLevel--
                 }
@@ -483,28 +485,12 @@ function Get-M365DSCDRGComplexTypeToString
     $indent = ''
     $indent = '    ' * ($IndentLevel -1)
 
-    if ($key -in $ComplexTypeMapping.Name)
+    if ($key -in $ComplexTypeMapping.Name -and -not $currentProperty.EndsWith("`r`n"))
     {
         $currentProperty += "`r`n"
     }
 
     $currentProperty += "$indent}"
-    <#
-    if ($IsArray -or $IndentLevel -gt 4)
-    {
-        $currentProperty += "`r`n"
-    }
-    #>
-
-    #Indenting last parenthesis when the cim instance is an array
-    <#
-    if ($IndentLevel -eq 5)
-    {
-        $indent = '    ' * ($IndentLevel -2)
-        $currentProperty += $indent
-    }
-    #>
-
     $emptyCIM = $currentProperty.Replace(' ', '').Replace("`r`n", '')
     if ($emptyCIM -eq "MSFT_$CIMInstanceName{}")
     {